Python/Project/fastapi-keycloak/requirements.txt
python-jose[cryptography]
fastapi[standard]
psycopg2-binary
fastapi-sso
sqlalchemy
requests
alembic
pandas
regex